:sparkles: feat(lab-java): TODO & tests for java lab filled

+++ src/main/java/com/code/a/sheep/codeasheep/domain/Hit.java
@@ -14,11 +14,11 @@ import lombok.Data;
 @Data
 @Builder
 public class Hit implements Comparable<Hit> {
-    private final float score;
+    private final double score;
     private final Document document;
 
     @Override
     public int compareTo(Hit o) {
-        return Float.compare(score, o.getScore());
+        return Double.compare(score, o.getScore());
     }
 }
